+++ core/kernel/src/feature_manifest_parsing.rs
