+++ core/kernel/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod c_interface;
-mod feature_manifest;
+mod feature_manifest_parsing;
 pub mod kernel;
 #[allow(unused_imports)]
 mod feature_manifest_generated;
